Quote field names when generating models
TypeScript only supports a subset of valid JSON keys as property names, but allows the use of quoted strings as properties.

Currently this means the typescript-fetch codegen will generate TypeScript code that can't compile if it's given a Swagger Definition that contains keys that aren't valid TypeScript property names. In my case, I was generating from an OData source, where there's special keys that start with '@'.

This small change fixes the problem.